changes of orderid

+++ OtpAPI/Controllers/V1Controller.cs
@@ -402,7 +402,7 @@ namespace OtpAPI.Controllers
                 {
                     if (Convert.ToInt16(insertUpdateOrder.OrderId) == 0)
                     {
-                        insertUpdateOrder.OrderId = (int)_otpBAL.GetOrderId();
+                        insertUpdateOrder.OrderId = _otpBAL.GetOrderId();
                     }
                     var results = _otpBAL.InsertOrder(insertUpdateOrder);
                     return Ok(results);
